backend/requirements.txt
Django==5.1.*
djangorestframework==3.15.*
django-filter==24.*
psycopg2-binary==2.9.*
python-dotenv==1.*
django-cors-headers==4.*
openai==1.*
dj-database-url==2.*
whitenoise==6.*
gunicorn==22.*
